Move functions from utils.ts closer to searchable-select component.
