+++ src/page-modules/contact/components/input/description-modal.tsx
@@ -43,7 +43,7 @@ const DescriptionModal = ({
       document.removeEventListener('keydown', handleEscapeOrClickOutside);
       document.removeEventListener('mousedown', handleEscapeOrClickOutside);
     };
-  });
+  }, []);
 
   if (!isModalOpen) return null;
