+++ src/page-modules/contact/travel-guarantee/forms/refundTaxiForm.tsx
@@ -11,7 +11,11 @@ import { Checkbox } from '../../components/input/checkbox';
 import { Textarea } from '../../components/input/textarea';
 import { FileInput } from '../../components/input/file';
 import SearchableSelect from '../../components/input/searchable-select';
-import { getLineOptions, getStopOptions } from '../../utils';
+import {
+  getLineOptions,
+  getStopOptions,
+} from '../../components/input/searchable-select/utils';
+
 type RefundTaxiFormProps = {
   state: { context: ContextProps };
   send: (event: typeof TravelGuaranteeFormEvents) => void;
