+++ src/page-modules/contact/utils.ts
@@ -1,7 +1,5 @@
 import { TransportModeType } from '@atb-as/config-specs';
 import { Line } from '.';
-import { Option } from './components/input/searchable-select';
-import { ComponentText, Language } from '@atb/translations';
 
 export const shouldShowContactPage = (): boolean => {
   return process.env.NEXT_PUBLIC_CONTACT_API_URL ? true : false;
@@ -75,22 +73,4 @@ export const setLineAndResetStops = (context: any, line: Line | undefined) => {
       toStop: [],
     },
   };
-};
-
-export const getLineOptions = (lines: Line[]): Option<Line>[] => {
-  return lines.map((line) => ({
-    id: line.id,
-    name: line.publicCode ? `${line.publicCode} - ${line.name}` : line.name,
-    value: line,
-  }));
-};
-
-export const getStopOptions = (
-  stops: Line['quays'],
-): Option<Line['quays'][0]>[] => {
-  return stops.map((stop) => ({
-    id: stop.id,
-    name: stop.name,
-    value: stop,
-  }));
 };
